Fix more connectivity/WiFi issues.
I was still having periodic connectivity issues that I
(finally) nailed down to a shoddy Intel Wireless firmware
issue. Add a to the README about that.

I also had issues with the NICs pulling up cleanly on a
scheduled restart, which I fixed by hacking the interfaces
file a little. I also removed the 172.x subnet since I
don't really have a use for it (I had originally planned
to isolate the WiFi into two separate network slices, but
the WiFI doesn't support multiple SSIDs)..

Lastly, alter the udev file a little as systemd-journald
was logging some messages about a comment appearing in
the middle of the line, which is supposedly not allowed.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>